* gitignore был добавлен для c++ приложений * Arduino - пофикшена компиляция * cxx-api - изменен путь для результатов * cxx-app - написан примитивный пример (пока не компилируется)

+++ code/Arduino/Constants.cpp
@@ -32,11 +32,11 @@ const char Constants::commands_delimetr = ';';
 
 const uint8_t Constants::min_speed = 0;
 const uint8_t Constants::max_speed = 255;
-const uint8_t Constants::servo_delay = 20;
+const uint16_t Constants::servo_delay = 20;
 
 const uint8_t Constants::countDistanceSensors = 5;
 const uint8_t Constants::countLineSensors = 5;
-const uint8_t Constants::minimalLineBound = 270;
+const uint16_t Constants::minimalLineBound = 270;
 
 Constants::Constants()
 {
